feat: access restrictions

+++ routes/api.php
@@ -36,7 +36,7 @@ Route::post('/flw-hook', [WebhooksManager::class, 'flwWebhook']);
 
 Route::group(['prefix' => 'v1', 'middleware' => ['cors', 'json.response']], function () {
 
-    
+
 
     Route::group(['prefix' => 'user'], function () {
 
@@ -64,47 +64,47 @@ Route::group(['prefix' => 'v1', 'middleware' => ['cors', 'json.response']], func
 
             Route::group(['prefix' => 'category', 'middleware' => []], function () {
                 Route::get('/', [CategoryManager::class, 'index']);
-                Route::post('/', [CategoryManager::class, 'store']);
+                Route::post('/', [CategoryManager::class, 'store'])->middleware(["auth.admin.access"]);
                 Route::get('/{category}', [CategoryManager::class, 'view']);
-                Route::patch('/{category}', [CategoryManager::class, 'update']);
-                Route::delete('/{category}', [CategoryManager::class, 'delete']);
+                Route::patch('/{category}', [CategoryManager::class, 'update'])->middleware(["auth.learner.access"]);
+                Route::delete('/{category}', [CategoryManager::class, 'delete'])->middleware(["auth.learner.access"]);
             });
 
             Route::group(['prefix' => 'course', 'middleware' => []], function () {
                 Route::get('/', [CourseManager::class, 'index']);
-                Route::post('/', [CourseManager::class, 'store']);
+                Route::post('/', [CourseManager::class, 'store'])->middleware(["auth.admin.access"]);
                 Route::get('/{course}', [CourseManager::class, 'view']);
-                Route::get('/{course}/subscribe', [CourseManager::class, 'subscribe']);
-                Route::post('/{course}', [CourseManager::class, 'update']);
-                Route::delete('/{course}', [CourseManager::class, 'delete']);            
+                Route::get('/{course}/subscribe', [CourseManager::class, 'subscribe'])->middleware(["auth.learner.access"]);;
+                Route::post('/{course}', [CourseManager::class, 'update'])->middleware(["auth.admin.access"]);
+                Route::delete('/{course}', [CourseManager::class, 'delete'])->middleware(["auth.admin.access"]);
             });
 
             Route::group(['prefix' => 'lesson', 'middleware' => []], function () {
                 Route::get('/', [LessonManager::class, 'index']);
-                Route::post('/{course}', [LessonManager::class, 'store']);
+                Route::post('/{course}', [LessonManager::class, 'store'])->middleware(["auth.admin.access"]);
                 Route::post('/{course}/list', [LessonManager::class, 'storeList']);
                 Route::get('/{lesson}', [LessonManager::class, 'view']);
-                Route::get('/{lesson}/seen', [LessonManager::class, 'seen']);
-                Route::post('/{lesson}', [LessonManager::class, 'update']);
-                Route::delete('/{lesson}', [LessonManager::class, 'delete']);
+                Route::get('/{lesson}/seen', [LessonManager::class, 'seen'])->middleware(["auth.learner.access"]);
+                Route::post('/{lesson}', [LessonManager::class, 'update'])->middleware(["auth.admin.access"]);
+                Route::delete('/{lesson}', [LessonManager::class, 'delete'])->middleware(["auth.admin.access"]);
             });
 
             Route::group(['prefix' => 'transaction', 'middleware' => []], function () {
                 Route::get('/', [TransactionManager::class, 'index']);
                 Route::get('/all', [TransactionManager::class, 'indexAll']);
                 Route::get('/{transaction}', [TransactionManager::class, 'view']);
-                Route::post('/{transaction}', [TransactionManager::class, 'update']);
-                Route::delete('/{transaction}', [TransactionManager::class, 'delete']);
-                Route::get('/{transaction}/payout', [TransactionManager::class, 'makePayout']);          
+                Route::post('/{transaction}', [TransactionManager::class, 'update'])->middleware(["auth.admin.access"]);
+                Route::delete('/{transaction}', [TransactionManager::class, 'delete'])->middleware(["auth.admin.access"]);
+                Route::get('/{transaction}/payout', [TransactionManager::class, 'makePayout'])->middleware(["auth.admin.access"]);
             });
 
             Route::group(['prefix' => 'assignment', 'middleware' => []], function () {
                 Route::get('/', [AssignmentManager::class, 'index']);
-                Route::post('/', [AssignmentManager::class, 'store']);
+                Route::post('/', [AssignmentManager::class, 'store'])->middleware(["auth.admin.access"]);
                 Route::get('/{assignment}', [AssignmentManager::class, 'view']);
-                Route::post('/{assignment}/submit', [AssignmentManager::class, 'submit']);
-                Route::patch('/{assignment}', [AssignmentManager::class, 'update']);
-                Route::delete('/{assignment}', [AssignmentManager::class, 'delete']);
+                Route::post('/{assignment}/submit', [AssignmentManager::class, 'submit'])->middleware(["auth.learner.access"]);
+                Route::patch('/{assignment}', [AssignmentManager::class, 'update'])->middleware(["auth.admin.access"]);
+                Route::delete('/{assignment}', [AssignmentManager::class, 'delete'])->middleware(["auth.admin.access"]);
             });
 
             Route::group(['prefix' => 'attachment', 'middleware' => []], function () {
@@ -126,8 +126,8 @@ Route::group(['prefix' => 'v1', 'middleware' => ['cors', 'json.response']], func
             Route::group(['prefix' => 'banking', 'middleware' => []], function () {
                 Route::group(['prefix' => 'paystack', 'middleware' => []], function () {
                     Route::get('/listbank', [PaystackManager::class, 'allBanks']);
-                    Route::get('/verify-account', [PaystackManager::class, 'verifyAccount']);          
-                    Route::post('/initialize-transaction', [PaystackManager::class, 'initializeTransaction']);          
+                    Route::get('/verify-account', [PaystackManager::class, 'verifyAccount']);
+                    Route::post('/initialize-transaction', [PaystackManager::class, 'initializeTransaction']);
                 });
             });
 
@@ -136,7 +136,7 @@ Route::group(['prefix' => 'v1', 'middleware' => ['cors', 'json.response']], func
                 Route::post('/', [QuizManager::class, 'store']);
                 Route::get('/{quiz}', [QuizManager::class, 'view']);
                 Route::patch('/{quiz}', [QuizManager::class, 'update']);
-                Route::post('/{quiz}/submit', [QuizManager::class, 'submit']);
+                Route::post('/{quiz}/submit', [QuizManager::class, 'submit'])->middleware(["auth.learner.access"]);
                 Route::delete('/{quiz}', [QuizManager::class, 'delete']);
             });
 
@@ -147,14 +147,12 @@ Route::group(['prefix' => 'v1', 'middleware' => ['cors', 'json.response']], func
                 Route::patch('/{question}', [QuestionManager::class, 'update']);
                 Route::delete('/{question}', [QuestionManager::class, 'delete']);
             });
-            
 
-            
+
+
 
 
             Route::get('/logout', [AuthManager::class, 'logout']);
         });
-
-        
     });
 });
